+++ HISTORY.md
@@ -11,6 +11,7 @@
 
 ### 2021-02-09
 * lists: refactored matrix ops to avoid JS stack overflows
+* objects: fixed internal migration for "transpose" block
 
 ### 2021-02-08
 * lists, objects, threads: new RESHAPE primitive